shopping cart service added along with unit tests

+++ Checkout/Contracts/IProduct.cs
@@ -6,6 +6,6 @@ namespace Checkout.Contracts
     {
         string SKU { get; set; }
 
-        double UnitPrice { get; set; }
+        decimal UnitPrice { get; set; }
     }
 }
